+++ app/src/main/java/com/paveynganpi/snapshare/ui/LoginActivity.java
@@ -23,6 +23,7 @@ import com.parse.ParseTwitterUtils;
 import com.parse.ParseUser;
 import com.parse.SaveCallback;
 import com.parse.twitter.Twitter;
+import com.paveynganpi.snapshare.POJO.TwitterFolloweePojo;
 import com.paveynganpi.snapshare.POJO.TwitterUserpojo;
 import com.paveynganpi.snapshare.R;
 import com.paveynganpi.snapshare.SnapShareApplication;
@@ -96,13 +97,13 @@ public class LoginActivity extends ActionBarActivity {
                             dialog.show();//show the dialog
                         } else if (user.isNew()) {
                             SnapShareApplication.updateParseInstallation(user);
-                            mCurrentUser =  ParseUser.getCurrentUser();
+                            mCurrentUser = ParseUser.getCurrentUser();
 
-//                            GetTwitterUserFolloweeIds getTwitterUserFolloweeIds = new GetTwitterUserFolloweeIds();
-//                            getTwitterUserFolloweeIds.execute();
+                            GetTwitterUserFolloweeIds getTwitterUserFolloweeIds = new GetTwitterUserFolloweeIds();
+                            getTwitterUserFolloweeIds.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
 
                             GetTwitterUserDataTask getTwitterUserDataTask = new GetTwitterUserDataTask();
-                            getTwitterUserDataTask.execute();
+                            getTwitterUserDataTask.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
 
                             user.setUsername("aaatest");
                             user.saveInBackground(new SaveCallback() {
@@ -129,15 +130,8 @@ public class LoginActivity extends ActionBarActivity {
                             intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
                             startActivity(intent);
                         } else {
+
                             SnapShareApplication.updateParseInstallation(user);
-                            mCurrentUser =  ParseUser.getCurrentUser();
-
-                            GetTwitterUserDataTask getTwitterUserDataTask = new GetTwitterUserDataTask();
-                            getTwitterUserDataTask.execute();
-
-//                            GetTwitterUserFolloweeIds getTwitterUserFolloweeIds = new GetTwitterUserFolloweeIds();
-//                            getTwitterUserFolloweeIds.execute();
-
 
                             Intent intent = new Intent(LoginActivity.this, MainActivity.class);
                             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
@@ -148,61 +142,6 @@ public class LoginActivity extends ActionBarActivity {
                 });
             }
 
-//                //get their various strings
-//                String username = mUsername.getText().toString();
-//                String password = mPassword.getText().toString();
-//
-//                //remove extra white spaces
-//                username.trim();
-//                password.trim();
-//
-//                //if user leaves any text fields blank
-//                if (username.isEmpty() || password.isEmpty()) {
-//                    AlertDialog.Builder builder = new AlertDialog.Builder(LoginActivity.this);
-//                    builder.setMessage(R.string.login_error_message);//creates a dialog with this message
-//                    builder.setTitle(R.string.login_error_title);
-//                    builder.setPositiveButton(android.R.string.ok, null);//creates a button to dismiss the dialog
-//
-//                    AlertDialog dialog = builder.create();//create a dialog
-//                    dialog.show();//show the dialog
-//                }
-//                //the user put in good data, logging in
-//                else {
-//                    //login
-//                    setProgressBarIndeterminateVisibility(true);//show the progress bar
-//
-//                    ParseUser.logInInBackground(username, password, new LogInCallback() {
-//                        @Override
-//                        public void done(ParseUser user, ParseException e) {
-//                            setProgressBarIndeterminateVisibility(false);//remove progress bar after we have response from parse
-//                            if (e == null) {
-//                                //success
-//
-//                                SnapShareApplication.updateParseInstallation(user);
-//
-//                                Intent intent = new Intent(LoginActivity.this, MainActivity.class);
-//                                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-//                                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
-//                                startActivity(intent);
-//
-//                            } else {
-//
-//                                AlertDialog.Builder builder = new AlertDialog.Builder(LoginActivity.this);
-//                                builder.setMessage(R.string.login_error_message);//creates a dialog with this message
-//                                builder.setTitle(R.string.login_error_title);
-//                                builder.setPositiveButton(android.R.string.ok, null);//creates a button to dismiss the dialog
-//
-//                                AlertDialog dialog = builder.create();//create a dialog
-//                                dialog.show();//show the dialog
-//
-//                            }
-//
-//                        }
-//                    });
-//
-//                }
-//
-//            }
         });
 
     }
@@ -273,61 +212,61 @@ public class LoginActivity extends ActionBarActivity {
         }
     }
 
-//    public class GetTwitterUserFolloweeIds extends AsyncTask<Object, Void, TwitterFolloweePojo> {
-//
-//        public StringBuilder sb = new StringBuilder();
-//        private Twitter currentTwitterUser = ParseTwitterUtils.getTwitter();
-//        public TwitterFolloweePojo twitterFolloweePojo;
-//
-//        @Override
-//        protected TwitterFolloweePojo doInBackground(Object... arg0) {
-//            HttpClient client = new DefaultHttpClient();
-//            HttpGet verifyGet = new HttpGet(
-//                    "https://api.twitter.com/1.1/friends/ids.json?cursor=-1&screen_name="
-//                            + currentTwitterUser.getScreenName() + "&stringify_ids=true&count=5000");
-//            currentTwitterUser.signRequest(verifyGet);
-//            try {
-//                HttpResponse response = client.execute(verifyGet);
-//
-//                //gets response body from response object
-//                BufferedReader reader =
-//                        new BufferedReader(new InputStreamReader(response.getEntity().getContent()), 65728);
-//                String line = null;
-//                while ((line = reader.readLine()) != null) {
-//                    sb.append(line);
-//                }
-//
-//                ObjectMapper mapper = new ObjectMapper();
-//                mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-//                twitterFolloweePojo = mapper.readValue(sb.toString(), TwitterFolloweePojo.class);
-//
-//            } catch (IOException e) {
-//                e.printStackTrace();
-//            } catch (Exception e) {
-//                e.printStackTrace();
-//            }
-//            return twitterFolloweePojo;
-//        }
-//
-//        @Override
-//        protected void onPostExecute(final TwitterFolloweePojo twitterFolloweePojo) {
-//            super.onPostExecute(twitterFolloweePojo);
-//            Log.d("post execute ee", "post execute works");
-//
-//            mCurrentUser.put("followeeIds", twitterFolloweePojo.getFolloweeIds());
-//            mCurrentUser.saveInBackground(new SaveCallback() {
-//                @Override
-//                public void done(ParseException e) {
-//                    if (e == null) {
-//                        Log.d("post execute", "successfully saved user in parse");
-//
-//                    } else {
-//                        Log.d("post execute", "error saving user in parse " + e.getMessage());
-//                    }
-//                }
-//            });
-//        }
-//    }
+    public class GetTwitterUserFolloweeIds extends AsyncTask<Object, Void, TwitterFolloweePojo> {
+
+        public StringBuilder sb = new StringBuilder();
+        private Twitter currentTwitterUser = ParseTwitterUtils.getTwitter();
+        public TwitterFolloweePojo twitterFolloweePojo;
+
+        @Override
+        protected TwitterFolloweePojo doInBackground(Object... arg0) {
+            HttpClient client = new DefaultHttpClient();
+            HttpGet verifyGet = new HttpGet(
+                    "https://api.twitter.com/1.1/friends/ids.json?cursor=-1&screen_name="
+                            + currentTwitterUser.getScreenName() + "&stringify_ids=true&count=5000");
+            currentTwitterUser.signRequest(verifyGet);
+            try {
+                HttpResponse response = client.execute(verifyGet);
+
+                //gets response body from response object
+                BufferedReader reader =
+                        new BufferedReader(new InputStreamReader(response.getEntity().getContent()), 65728);
+                String line = null;
+                while ((line = reader.readLine()) != null) {
+                    sb.append(line);
+                }
+
+                ObjectMapper mapper = new ObjectMapper();
+                mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+                twitterFolloweePojo = mapper.readValue(sb.toString(), TwitterFolloweePojo.class);
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            return twitterFolloweePojo;
+        }
+
+        @Override
+        protected void onPostExecute(final TwitterFolloweePojo twitterFolloweePojo) {
+            super.onPostExecute(twitterFolloweePojo);
+            Log.d("post execute ee", "post execute works");
+
+            mCurrentUser.put("followeeIds", twitterFolloweePojo.getFolloweeIds());
+            mCurrentUser.saveInBackground(new SaveCallback() {
+                @Override
+                public void done(ParseException e) {
+                    if (e == null) {
+                        Log.d("post execute", "successfully saved user in parse");
+
+                    } else {
+                        Log.d("post execute", "error saving user in parse " + e.getMessage());
+                    }
+                }
+            });
+        }
+    }
 
     private void navigateToLogin() {
         //start the loginActivity
